input: synaptics: s3320: Ensure I2C bus is active inside ISR thread
When the device is suspended, irq_wake is apparently not enough to ensure
that the I2C bus is active while an interrupt is processed.

In order to prevent gestures from failing during power collapse, check if
the I2C bus is active inside the ISR thread. If it isn't, wait until it
becomes active before processing the interrupt (typically a ~10ms wait).

This fixes off-screen gestures not working sometimes.

Signed-off-by: Sultanxda <sultanxda@gmail.com>